+++ RELEASES.md
@@ -3,6 +3,8 @@ Use Element.namespaceSeperator in the construction of fully qualified names.
 
 Remove unnecessary template parameters to methods that accept behaviour callbacks.
 
+Add documentation for Transition.source and Transition.target.
+
 ## Version 5.6.0
 Fix a possible issue where deep history semantics were not properly propigated through the state machine model initialisation.